requirements.txt
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pdf2image>=1.16.0
watchdog>=3.0.0
python-dotenv>=1.0.0
flask>=3.0.0

